Move blockchain related stuff into separate folders
Pull, watch and generate functionality for each blockchain now located
in own directory.
