+++ src/blockchains/eos/eos-js-gen.js
